+++ tests/testthat/test-gisco-address-api.R
@@ -1,9 +1,5 @@
-test_that("Address API returns NULL when offline", {
-  skip_on_cran()
-  skip_if_gisco_offline()
-  local_mocked_bindings(is_online_fun = function(...) {
-    FALSE
-  })
+test_that("Address API returns NULL when connection fails", {
+  local_mocked_bindings(gisco_req_perform = mock_connection_failure)
 
   expect_snapshot(fend <- gisco_address_api_bbox())
   expect_null(fend)